requirements.txt
altgraph==0.17
astroid==2.4.2
autopep8==1.5.3
bleach==3.1.5
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cssselect==1.1.0
decorator==4.4.2
dicttoxml==1.7.4
docutils==0.16
future==0.18.2
geocoder==1.38.1
geoip2==3.0.0
idna==2.9
ip2geotools==0.1.5
IP2Location==8.4.1
isort==4.3.21
keyring==21.2.1
lazy-object-proxy==1.5.0
lxml==4.5.1
maxminddb==1.5.4
mccabe==0.6.1
packaging==20.4
pefile==2019.4.18
Pillow==7.1.2
pip-review==1.1.0
pkginfo==1.5.0.1
pycodestyle==2.6.0
Pygments==2.6.1
PyInstaller==3.6
pylint==2.5.3
pyparsing==2.4.7
pyquery==1.4.1
pywin32==228
pywin32-ctypes==0.2.0
ratelim==0.1.6
readme-renderer==26.0
requests==2.24.0
requests-toolbelt==0.9.1
selenium==3.141.0
six==1.15.0
toml==0.10.1
tqdm==4.46.1
twine==3.1.1
typed-ast==1.4.1
typing==3.7.4.1
urllib3==1.25.9
webencodings==0.5.1
wrapt==1.12.1